renamed player pages
