+++ ui/pages/player_info_page.py
